+++ pytket/binders/passes.cpp
@@ -58,7 +58,7 @@ Transforms::TwoQbFidelities get_fidelities(const py::kwargs &kwargs) {
     } else if (kwargstr == "ZZMax_fidelity") {
       fid.ZZMax_fidelity = py::cast<double>(kwarg.second);
     } else if (kwargstr == "ZZPhase_fidelity") {
-      fid.ZZPhase_fidelity = py::cast<Func>(kwarg.second);
+      fid.ZZPhase_fidelity = py::cast<std::variant<double, Func>>(kwarg.second);
     } else {
       throw py::type_error(
           "got an unexpected keyword argument '" + kwargstr + "'");
@@ -405,9 +405,9 @@ PYBIND11_MODULE(passes, m) {
       "`ZZPhase_fidelity`. If provided, the `CX` and `ZZMax` fidelities "
       "must be given by a single floating point fidelity. The `ZZPhase` "
       "fidelity is given as a lambda float -> float, mapping a ZZPhase "
-      "angle parameter to its fidelity. These parameters will be used "
-      "to return the optimal decomposition of each TK2 gate, taking "
-      "noise into consideration.\n\n"
+      "angle parameter to its fidelity, or by a single float. These parameters "
+      "will be used to return the optimal decomposition of each TK2 gate, "
+      "taking noise into consideration.\n\n"
       "If no fidelities are provided, the TK2 gates will be decomposed "
       "exactly using CX gates. For equal fidelities, ZZPhase will be prefered "
       "over ZZMax and CX if the decomposition results in fewer two-qubit "