+++ pytket/docs/changelog.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+Features:
+
+* DecomposeTK2 pass and transform can now accept a float for ZZPhase_fidelity.
+* DecomposeTK2 pass now has a json representation when it contains no functions.
+
 1.32.0 (September 2024)
 -----------------------